Moved orderState to module
