+++ core/modules/order/store/mutations.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import { MutationTree } from 'vuex'
 import * as types from './mutation-types'
 import * as entities from '@vue-storefront/store/lib/entities'
-import OrderState from './types/OrderState'
+import OrderState from '../types/OrderState'
 import rootStore from '@vue-storefront/store'
 
 const mutations: MutationTree<OrderState> = {
